Add files via upload
This script allow the user to rename ALL files in a specific folder to a new name plus a sequential number.  This can be used on large folders of pictures, etc.  The user must input the path of the folder and the desired new name of the files.  The script transfers over the current file extensions of all files and does NOT change the name of any folders or files with no extension in the target folder.